refactor: split Sub.hs
